[ardpyGUI]
finished the first layout.
to handle the pins/events correctly, lists that holds all objects by type where created so we can access the relative object based on the pin numbers.

Function dw() distinguish between regular active/deactive pins and pwn active pins. Activate and deactivate changing the relief parameter on the button object. If PWM active it reads the corresponding scale value.

Function act() is call at PWM checkbox todggling, it updates the array lstPWM that is used for dw() to know if pin is acting as regular pin or PWM.